--- client/connexion.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/wetteren_logo.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Page de connexion</title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/connexion.jsx"></script>
-  </body>
-</html>
