+++ client/package.json
@@ -34,6 +34,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
     "postcss": "^8.4.35",
+    "react-router-dom": "^6.22.3",
     "tailwindcss": "^3.4.1",
     "vite": "^5.1.4"
   }
