+++ client/src/pages/Layout.jsx
@@ -1,0 +1,14 @@
+import Header from "../components/ui/Header/Header.jsx"
+import Footer from "../components/ui/Footer/Footer.jsx"
+
+//Pour afficher le header et footer sur toutes les pages
+function Layout(){
+    return (
+        <>
+            <Header />
+            <Footer />
+        </>
+    );
+}
+
+export default Layout
