+++ client/src/pages/NoPage.jsx
@@ -1,0 +1,7 @@
+function NoPage(){
+    return (
+        <p>Aucune page n'a pu être charger</p>
+    )
+}
+
+export default NoPage;
