+++ client/src/pages/Register.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import axios from "axios";
 
-const App = () => {
+const Register = () => {
   const [name, setName] = useState("");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
@@ -60,7 +60,7 @@ const App = () => {
   );
 };
 
-export default App;
+export default Register;
